🎉 Implement error handling functions for improved response management
