+++ services/web/server/src/simcore_service_webserver/exception_handling/__init__.py
@@ -1,9 +1,12 @@
+from servicelib.aiohttp.web_exceptions_handling import (
+    create_error_context_from_request,
+    create_error_response,
+)
+
 from ._base import ExceptionHandlersMap, exception_handling_decorator
 from ._factory import (
     ExceptionToHttpErrorMap,
     HttpErrorInfo,
-    create_error_context_from_request,
-    create_error_response,
     to_exceptions_handlers_map,
 )
