+++ services/web/server/src/simcore_service_webserver/exception_handling/_factory.py
@@ -1,15 +1,16 @@
 import logging
-from typing import Any, NamedTuple, TypeAlias
+from typing import NamedTuple, TypeAlias
 
 from aiohttp import web
 from common_library.error_codes import create_error_code
-from common_library.json_serialization import json_dumps
 from common_library.logging.logging_errors import create_troubleshooting_log_kwargs
 from models_library.rest_error import ErrorGet
-from servicelib.aiohttp.rest_responses import safe_status_message
 from servicelib.aiohttp.web_exceptions_extension import get_all_aiohttp_http_exceptions
+from servicelib.aiohttp.web_exceptions_handling import (
+    create_error_context_from_request,
+    create_error_response,
+)
 from servicelib.status_codes_utils import (
-    get_code_display_name,
     is_5xx_server_error,
     is_error,
 )
@@ -37,26 +38,6 @@ class HttpErrorInfo(NamedTuple):
 
 
 ExceptionToHttpErrorMap: TypeAlias = dict[type[Exception], HttpErrorInfo]
-
-
-def create_error_context_from_request(request: web.Request) -> dict[str, Any]:
-    return {
-        "request": request,
-        "request.remote": f"{request.remote}",
-        "request.method": f"{request.method}",
-        "request.path": f"{request.path}",
-    }
-
-
-def create_error_response(error: ErrorGet, status_code: int) -> web.Response:
-    assert is_error(status_code), f"{status_code=} must be an error [{error=}]"  # nosec
-
-    return web.json_response(
-        data={"error": error.model_dump(exclude_unset=True, mode="json")},
-        dumps=json_dumps,
-        reason=safe_status_message(get_code_display_name(status_code)),
-        status=status_code,
-    )
 
 
 def create_exception_handler_from_http_info(
